feat: refactor AdLibPanel code to be more modular and use React hooks. Memoize as much as possible for better performance.
