+++ meteor/client/ui/Shelf/AdLibPanel.tsx
@@ -1,18 +1,15 @@
 import React, { useState, useCallback, useEffect } from 'react'
 import _ from 'underscore'
 import { Meteor } from 'meteor/meteor'
-import { Translated, useTracker } from '../../lib/ReactMeteorData/react-meteor-data'
-import { useTranslation, withTranslation } from 'react-i18next'
+import { useTracker } from '../../lib/ReactMeteorData/react-meteor-data'
+import { useTranslation } from 'react-i18next'
 import { Rundown, RundownId } from '../../../lib/collections/Rundowns'
 import { RundownPlaylist, RundownPlaylistCollectionUtil } from '../../../lib/collections/RundownPlaylists'
 import { DBSegment, Segment, SegmentId } from '../../../lib/collections/Segments'
 import { DBPart, PartId } from '../../../lib/collections/Parts'
 import { AdLibPiece, AdLibPieces } from '../../../lib/collections/AdLibPieces'
-import { AdLibListItem, IAdLibListItem } from './AdLibListItem'
+import { IAdLibListItem } from './AdLibListItem'
 import ClassNames from 'classnames'
-
-import { faTh, faList, faTimes } from '@fortawesome/free-solid-svg-icons'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 
 import { Spinner } from '../../lib/Spinner'
 import { ShowStyleBase } from '../../../lib/collections/ShowStyleBases'
@@ -34,7 +31,6 @@ import {
 	RundownBaselineAdLibItem,
 	RundownBaselineAdLibPieces,
 } from '../../../lib/collections/RundownBaselineAdLibPieces'
-import { Random } from 'meteor/random'
 import { literal, normalizeArray, unprotectString, protectString } from '../../../lib/lib'
 import { memoizedIsolatedAutorun } from '../../lib/reactiveData/reactiveDataHelper'
 import {
@@ -61,380 +57,8 @@ import { i18nTranslator } from '../i18n'
 import { getShelfFollowsOnAir, getShowHiddenSourceLayers } from '../../lib/localStorage'
 import { sortAdlibs } from '../../../lib/Rundown'
 import { PieceStatusCode } from '@sofie-automation/corelib/dist/dataModel/Piece'
-
-interface IListViewPropsHeader {
-	uiSegments: Array<AdlibSegmentUi>
-	onSelectAdLib?: (piece: IAdLibListItem) => void
-	onToggleAdLib?: (piece: IAdLibListItem, queue: boolean, e: KeyboardEvent, mode?: IBlueprintActionTriggerMode) => void
-	selectedPiece: BucketAdLibActionUi | BucketAdLibUi | IAdLibListItem | PieceUi | undefined
-	selectedSegment: AdlibSegmentUi | undefined
-	searchFilter: string | undefined
-	showStyleBase: ShowStyleBase
-	noSegments: boolean
-	filter: RundownLayoutFilter | undefined
-	rundownAdLibs?: Array<AdLibPieceUi>
-	playlist: RundownPlaylist
-	studio: Studio
-}
-
-interface IListViewStateHeader {
-	outputLayers: {
-		[key: string]: IOutputLayer
-	}
-	sourceLayers: {
-		[key: string]: ISourceLayer
-	}
-}
-
-/**
- * Applies a filter to an adLib to determine whether it matches filter criteria.
- * @param item AdLib to test against filter.
- * @param showStyleBase
- * @param uiSegments All segments to search for live segment.
- * @param filter Filter to match against.
- * @param searchFilter Text to try to match against adLib label.
- * @param uniquenessIds Set of uniquenessIds, for a given set only one adLib per uniquness Id will be matched by this filter.
- */
-export function matchFilter(
-	item: AdLibPieceUi,
-	showStyleBase: ShowStyleBase,
-	uiSegments: Array<AdlibSegmentUi>,
-	filter?: RundownLayoutFilterBase,
-	searchFilter?: string,
-	uniquenessIds?: Set<string>
-) {
-	if (!searchFilter && !filter) return true
-	const liveSegment = uiSegments.find((i) => i.isLive === true)
-	const uppercaseLabel = item.name.toUpperCase()
-	if (filter) {
-		// Filter currentSegment only
-		if (
-			filter.currentSegment === true &&
-			item.partId &&
-			((liveSegment && liveSegment.parts.find((i) => item.partId === i.part._id) === undefined) || !liveSegment)
-		) {
-			return false
-		}
-		// Filter out items that are not within outputLayerIds filter
-		if (
-			filter.outputLayerIds !== undefined &&
-			filter.outputLayerIds.length &&
-			filter.outputLayerIds.indexOf(item.outputLayerId) < 0
-		) {
-			return false
-		}
-		// Source layers
-		if (
-			filter.sourceLayerIds !== undefined &&
-			filter.sourceLayerIds.length &&
-			filter.sourceLayerIds.indexOf(item.sourceLayerId) < 0
-		) {
-			return false
-		}
-		// Source layer types
-		const sourceLayerType = showStyleBase.sourceLayers.find((i) => i._id === item.sourceLayerId)
-		if (
-			sourceLayerType &&
-			filter.sourceLayerTypes !== undefined &&
-			filter.sourceLayerTypes.length &&
-			filter.sourceLayerTypes.indexOf(sourceLayerType.type) < 0
-		) {
-			return false
-		}
-		// Item label needs at least one of the strings in the label array
-		if (
-			filter.label !== undefined &&
-			filter.label.length &&
-			filter.label.reduce((p, v) => {
-				return p || uppercaseLabel.indexOf(v.toUpperCase()) >= 0
-			}, false) === false
-		) {
-			return false
-		}
-		// Item tags needs to contain all of the strings in the tags array
-		if (
-			filter.tags !== undefined &&
-			filter.tags.length &&
-			filter.tags.reduce((p, v) => {
-				return p && item.tags !== undefined && item.tags.indexOf(v) >= 0
-			}, true) === false
-		) {
-			return false
-		}
-		// Hide duplicates
-		// Only the first adLib found with a given uniquenessId will be displayed if this option is enabled.
-		// Scope of the filter is determined by the scope of the uniquenessIds set (typically rundown-wide).
-		if (filter.hideDuplicates && uniquenessIds) {
-			const uniquenessId = item.uniquenessId || unprotectString(item._id)
-			if (uniquenessIds.has(uniquenessId)) {
-				return false
-			} else {
-				uniquenessIds.add(uniquenessId)
-			}
-		}
-	}
-	if (searchFilter) {
-		return uppercaseLabel.indexOf(searchFilter.trim().toUpperCase()) >= 0
-	} else {
-		return true
-	}
-}
-
-export function matchTags(item: AdLibPieceUi, tags?: string[]) {
-	if (
-		tags !== undefined &&
-		tags.reduce((p, v) => {
-			return p && item.tags !== undefined && item.tags.indexOf(v) >= 0
-		}, true) === false
-	) {
-		return false
-	}
-	return true
-}
-
-const AdLibListView = withTranslation()(
-	class AdLibListView extends React.Component<Translated<IListViewPropsHeader>, IListViewStateHeader> {
-		table: HTMLTableElement
-
-		constructor(props: Translated<IListViewPropsHeader>) {
-			super(props)
-
-			this.state = {
-				outputLayers: {},
-				sourceLayers: {},
-			}
-		}
-
-		static getDerivedStateFromProps(props: IListViewPropsHeader) {
-			const tOLayers: {
-				[key: string]: IOutputLayer
-			} = {}
-			const tSLayers: {
-				[key: string]: ISourceLayer
-			} = {}
-
-			if (props.showStyleBase && props.showStyleBase.outputLayers && props.showStyleBase.sourceLayers) {
-				props.showStyleBase.outputLayers.forEach((outputLayer) => {
-					tOLayers[outputLayer._id] = outputLayer
-				})
-				props.showStyleBase.sourceLayers.forEach((sourceLayer) => {
-					tSLayers[sourceLayer._id] = sourceLayer
-				})
-
-				return {
-					outputLayers: tOLayers,
-					sourceLayers: tSLayers,
-				}
-			}
-			return null
-		}
-
-		scrollToCurrentSegment() {
-			if (this.table.id && this.props.selectedSegment) {
-				// scroll to selected segment
-				const segmentSelector = `#${this.table.id} .adlib-panel__list-view__item__${this.props.selectedSegment._id}`
-				const segment: HTMLElement | null = document.querySelector(segmentSelector)
-				if (segment) {
-					this.table.scrollTo({
-						top: segment.offsetTop,
-						behavior: 'smooth',
-					})
-				}
-			}
-		}
-
-		componentDidMount() {
-			this.scrollToCurrentSegment()
-		}
-
-		componentDidUpdate(prevProps: IListViewPropsHeader) {
-			if (prevProps.selectedSegment !== this.props.selectedSegment) {
-				this.scrollToCurrentSegment()
-			}
-		}
-
-		renderRundownAdLibs(uniquenessIds: Set<string>) {
-			return (
-				<tbody className="adlib-panel__list-view__list__segment adlib-panel__list-view__item__rundown-baseline">
-					{this.props.rundownAdLibs &&
-						this.props.rundownAdLibs
-							.filter(
-								(item) =>
-									!item.isHidden &&
-									matchFilter(
-										item,
-										this.props.showStyleBase,
-										this.props.uiSegments,
-										this.props.filter,
-										this.props.searchFilter,
-										uniquenessIds
-									)
-							)
-							.map((adLibPiece: AdLibPieceUi) => (
-								<AdLibListItem
-									key={unprotectString(adLibPiece._id)}
-									piece={adLibPiece}
-									layer={adLibPiece.sourceLayer!}
-									studio={this.props.studio}
-									selected={
-										(this.props.selectedPiece &&
-											RundownUtils.isAdLibPiece(this.props.selectedPiece) &&
-											this.props.selectedPiece._id === adLibPiece._id) ||
-										false
-									}
-									disabled={adLibPiece.disabled ?? false}
-									onToggleAdLib={this.props.onToggleAdLib}
-									onSelectAdLib={this.props.onSelectAdLib}
-									playlist={this.props.playlist}
-								/>
-							))}
-				</tbody>
-			)
-		}
-
-		renderSegments(uniquenessIds: Set<string>) {
-			return this.props.uiSegments
-				.filter((a) => (this.props.filter ? (this.props.filter.currentSegment ? a.isLive : true) : true))
-				.map((segment) => {
-					const segmentAdLibs = segment.pieces.filter((item) =>
-						matchFilter(
-							item,
-							this.props.showStyleBase,
-							this.props.uiSegments,
-							this.props.filter,
-							this.props.searchFilter,
-							uniquenessIds
-						)
-					)
-
-					// only show the segment in the list if it's not hidden or it does contain some AdLibs
-					return !segment.isHidden || segment.pieces.length > 0 ? (
-						<tbody
-							key={unprotectString(segment._id)}
-							className={ClassNames(
-								'adlib-panel__list-view__list__segment',
-								'adlib-panel__list-view__item__' + segment._id,
-								{
-									live: segment.isLive,
-									next: segment.isNext && !segment.isLive,
-									past:
-										segment.parts.reduce((memo, item) => {
-											return item.timings?.startedPlayback && item.timings?.duration ? memo : false
-										}, true) === true,
-								}
-							)}
-						>
-							<tr className="adlib-panel__list-view__list__seg-header">
-								<td colSpan={4}>{segment.name}</td>
-							</tr>
-							{segmentAdLibs.map((adLibPiece: AdLibPieceUi) => (
-								<AdLibListItem
-									key={unprotectString(adLibPiece._id)}
-									piece={adLibPiece}
-									layer={adLibPiece.sourceLayer!}
-									studio={this.props.studio}
-									selected={
-										(this.props.selectedPiece &&
-											RundownUtils.isAdLibPiece(this.props.selectedPiece) &&
-											this.props.selectedPiece._id === adLibPiece._id) ||
-										false
-									}
-									disabled={adLibPiece.disabled ?? false}
-									onToggleAdLib={this.props.onToggleAdLib}
-									onSelectAdLib={this.props.onSelectAdLib}
-									playlist={this.props.playlist}
-								/>
-							))}
-						</tbody>
-					) : null
-				})
-		}
-
-		setTableRef = (el) => {
-			this.table = el
-		}
-
-		render() {
-			const uniquenessIds = new Set<string>()
-
-			return (
-				<div
-					className={ClassNames('adlib-panel__list-view__list', {
-						'adlib-panel__list-view__list--no-segments': this.props.noSegments,
-					})}
-				>
-					<table
-						id={'adlib-panel__list-view__table__' + Random.id()}
-						className="adlib-panel__list-view__list__table scroll-sink"
-						ref={this.setTableRef}
-					>
-						{this.renderRundownAdLibs(uniquenessIds)}
-						{this.renderSegments(uniquenessIds)}
-					</table>
-				</div>
-			)
-		}
-	}
-)
-
-interface IToolbarPropsHeader {
-	onFilterChange?: (newFilter: string | undefined) => void
-	noSegments?: boolean
-	searchFilter: string | undefined
-}
-
-export function AdLibPanelToolbar(props: IToolbarPropsHeader) {
-	const { t } = useTranslation()
-
-	function searchInputChanged(e?: React.ChangeEvent<HTMLInputElement>) {
-		const newValue = e?.target.value || ''
-		props.onFilterChange && typeof props.onFilterChange === 'function' && props.onFilterChange(newValue)
-	}
-
-	function searchInputKeyDown(e: React.KeyboardEvent<HTMLInputElement>) {
-		if (e.key === 'Escape' || e.key === 'Enter') {
-			document.querySelector('button')?.focus()
-		} else if (e.key.match(/^F\d+$/)) {
-			e.preventDefault()
-		}
-	}
-
-	function clearSearchInput() {
-		searchInputChanged()
-	}
-
-	return (
-		<div
-			className={ClassNames('adlib-panel__list-view__toolbar', {
-				'adlib-panel__list-view__toolbar--no-segments': props.noSegments,
-			})}
-		>
-			<div className="adlib-panel__list-view__toolbar__filter">
-				<input
-					className="adlib-panel__list-view__toolbar__filter__input"
-					type="text"
-					placeholder={t('Search...')}
-					onChange={searchInputChanged}
-					onKeyDown={searchInputKeyDown}
-					value={props.searchFilter || ''}
-				/>
-				{props.searchFilter !== '' && (
-					<div className="adlib-panel__list-view__toolbar__filter__clear" onClick={clearSearchInput}>
-						<FontAwesomeIcon icon={faTimes} />
-					</div>
-				)}
-			</div>
-			<div className="adlib-panel__list-view__toolbar__buttons" style={{ display: 'none' }}>
-				<button className="action-btn">
-					<FontAwesomeIcon icon={faList} />
-				</button>
-				<button className="action-btn">
-					<FontAwesomeIcon icon={faTh} />
-				</button>
-			</div>
-		</div>
-	)
-}
+import { AdLibPanelToolbar } from './AdLibPanelToolbar'
+import { AdLibListView } from './AdLibListView'
 
 export interface AdLibPieceUi extends AdLibPiece {
 	sourceLayer?: ISourceLayer
